requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langchain-huggingface
chromadb
pypdf
python-dotenv
sentence-transformers
rank_bm25
streamlit-oauth
